refactor: split types into a types.ts
